features/requirements.txt
pydantic>=2.0
python-dotenv
pandas>=2.0.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
click>=8.1.0
rich>=13.0.0
